👌 IMPROVE: Catch and log submissions
Instead of simply failing when one set of extras fails to produce correct inputs or submit a work chain, we now catch the error and report the failure in the submission loop of `submit_new_batch()`.

We also add a new `verbose` keyword argument. If this is set to `True` successful submissions are also reported.