✨ `FromGroupSubmissionController`: Add `filters`
Add a `filters` field for the `FromGroupSubmissionController` model,
which is passed to the `append()` call of the `Querybuilder` that
returns the nodes in the parent group based on the extras.

This allows users to define any filters supported by the AiiDA
`Querybuilder` when initiating the `FromGroupSubmissionController`,
which are then applied during the submission to filter which parent
nodes to create inputs for and submit.