👌 Always submit number of requested processes
Now that the `BaseSubmissionController` skips processes for which it is unable
to obtain a populated build or submit the process, the _actual_ number of
submitted process will be less than the currently available slots. This
becomes especially problematic when the number of failures is equal to the
number of available slots, in which case the submission controller basically
stops working.

Here we simply adapt the approach to loop over the `extras_to_run`, and only
stop in case the length of the `submitted` dictionary is equal or greater to
the number of available slots.