‼️ Prioritize the `order_by` approach to sorting submissions
Currently the `order_by` field of the `FromGroupSubmissionController` does not
actually work as intented for default `submit_new_batch` calls, since it will
sort the `extras_to_run` again after the extras "to submit" are obtained.
Moreover, the `get_all_extras_to_submit` method of the
`FromGroupSubmissionController` would return the extras as a set, instead of the
list obtained by the ordered query.

Here we set the `sort` input argument of the `submit_new_batch()` method to
`False` by default. This is technically a backwards-incompatible change,
although any impact on users should be minor since it only changes the order
of submission.