🐛 Allow specifying nested extras
Currently, any nested extras can not be used for the `BaseSubmissionController`.
Using e.g. `source.database` as one of the extras would fail, since when
adding the extras to the submitted work chain the controller attempts to
directly pass the `source.database` string as an extra key, which is not
accepted by AiiDA.

Here add the `get_extras_dict` utility function to convert the list of
(possibly nested) extras into a proper dictionary to pass to the
`set_extra_many` call on the work chain. This makes using nested extras for the
submission controllers possible.