🐛 Make `unique_extra_keys` field properly optional
The purpose of the `unique_extra_keys` field is that the user can define the keys of
the extras that track which calculations have been run upon instantiation of the
submission controller. However, it's also possible the user might override the
`get_extra_unique_keys()` method to fix these keys, as is done in the `add_in_batches`
example. In this case the user should not have to specify the `unique_extra_keys`,
but that currently doesn't work because we didn't set a default. This is fixed here.