Fix `submit_new_batch` when `verbose = True`
The `submit_new_batch` method fails if `verbose` is set to `True`, as it
refers to the `parent_group` to get the total number of submissions.
When using the `BaseSubmissionController`, this attribute is not
specified.
To fix the `verbose` option for the `BaseSubmissionController` and
`FromGroupSubmissionController`, the number of submissions is
defined by the number of `all extras to `submit`.