🐛 Fix default `unique_extra_keys` value
The current default value of `None` for the `unique_extra_keys` in the
`BaseSubmissionController` is not a valid approach for setting the default.
For `BaseSubmissionController`s, the user should still specify the unique extra keys,
since they can't be derived from a parent node. For the
`FromGroupSubmissionController`, it is sensible to use the `_aiida_hash` as a default
value, instead of `None`.